Move test data to a separate file
